Finish off E2E availability tests refactor by subsuming ViewWeek tests. Fix reused page issue by newing up new browserContext for certain tests.
